Finish profile service responsibility trim
