+++ src/Humans.Web/Controllers/ProfileController.cs
@@ -75,7 +75,6 @@ public class ProfileController(
     IApplicationDecisionService applicationDecisionService,
     IAccountDeletionService accountDeletionService,
     IMembershipCalculator membershipCalculator,
-    IHttpClientFactory httpClientFactory,
     SignInManager<User> signInManager,
     IOptions<GoogleWorkspaceOptions> googleWorkspaceOptions) : HumansControllerBase(userService)
 {
@@ -84,8 +83,6 @@ public class ProfileController(
     private readonly GoogleWorkspaceOptions _googleWorkspaceOptions = googleWorkspaceOptions.Value;
 
     private const int MaxProfilePictureUploadBytes = 20 * 1024 * 1024; // 20MB upload limit
-    private const int MaxGooglePhotoDownloadBytes = 20 * 1024 * 1024; // 20MB hard ceiling for Google avatar fetch
-    private const string GoogleAvatarHttpClientName = "GoogleAvatar";
     private static readonly HashSet<string> AllowedImageContentTypes = new(StringComparer.OrdinalIgnoreCase)
     {
         "image/jpeg",
@@ -222,15 +219,12 @@ public class ProfileController(
         var preferredShiftTags = userShiftView.TagPreferences
             .Select(p => new ShiftTagPreferenceSummary(p.ShiftTagId, p.ShiftTag?.Name ?? string.Empty))
             .ToList();
-        var externalLogins = await userManager.GetLoginsAsync(user);
-
         var viewModel = ProfileEditViewModelBuilder.Build(
             info,
             applications,
             allShiftTags,
             preferredShiftTags,
             preview,
-            externalLogins.Any(l => string.Equals(l.LoginProvider, "Google", StringComparison.OrdinalIgnoreCase)),
             p => Url.Action(nameof(Picture), new { id = p.Id, v = p.UpdatedAt.ToUnixTimeTicks() }));
 
         ViewData["GoogleMapsApiKey"] = configuration.GetRequiredSetting(configRegistry, "GoogleMaps:ApiKey", "Google Maps", isSensitive: true);
@@ -1652,130 +1646,6 @@ public class ProfileController(
         return File(result.Value.Data, result.Value.ContentType);
     }
 
-    [HttpPost("Me/ImportGooglePhoto")]
-    [ValidateAntiForgeryToken]
-    public async Task<IActionResult> ImportGooglePhoto(CancellationToken ct)
-    {
-        var user = await userManager.GetUserAsync(User);
-        if (user is null)
-        {
-            return NotFound();
-        }
-
-        var externalLogins = await userManager.GetLoginsAsync(user);
-        if (!HasGoogleAvatarSource(user, externalLogins))
-        {
-            SetError(localizer["Profile_ImportGooglePhoto_Unavailable"].Value);
-            return RedirectToAction(nameof(Edit));
-        }
-
-        var info = await _userService.GetUserInfoAsync(user.Id, ct);
-        if (info?.Profile is null)
-        {
-            SetError(localizer["Profile_ImportGooglePhoto_NoProfile"].Value);
-            return RedirectToAction(nameof(Edit));
-        }
-        if (info.Profile.HasCustomPicture)
-        {
-            SetError(localizer["Profile_ImportGooglePhoto_AlreadyHasCustom"].Value);
-            return RedirectToAction(nameof(Edit));
-        }
-
-        if (!TryGetTrustedGoogleAvatarUri(user, out var pictureUri))
-        {
-            SetError(localizer["Profile_ImportGooglePhoto_NotGoogleUrl"].Value);
-            return RedirectToAction(nameof(Edit));
-        }
-
-        var rawBytes = await FetchGoogleAvatarBytesAsync(pictureUri, user.Id, ct);
-        if (rawBytes is null)
-        {
-            return RedirectToAction(nameof(Edit));
-        }
-
-        var resized = Helpers.ProfilePictureProcessor.ResizeProfilePicture(rawBytes, logger);
-        if (resized is null)
-        {
-            SetError(localizer["Profile_ImportGooglePhoto_InvalidFormat"].Value);
-            return RedirectToAction(nameof(Edit));
-        }
-
-        await profilePictureService.SetProfilePictureAsync(user.Id, resized.Value.Data, resized.Value.ContentType, ct);
-
-        logger.LogInformation("Imported Google avatar for user {UserId}", user.Id);
-        SetSuccess(localizer["Profile_ImportGooglePhoto_Success"].Value);
-        return RedirectToAction(nameof(Edit));
-    }
-
-
-    private static bool HasGoogleAvatarSource(User user, IEnumerable<UserLoginInfo> externalLogins) =>
-        !string.IsNullOrEmpty(user.ProfilePictureUrl)
-        && externalLogins.Any(l => string.Equals(l.LoginProvider, "Google", StringComparison.OrdinalIgnoreCase));
-
-    private bool TryGetTrustedGoogleAvatarUri(User user, out Uri pictureUri)
-    {
-        if (Uri.TryCreate(user.ProfilePictureUrl, UriKind.Absolute, out pictureUri!)
-            && string.Equals(pictureUri.Scheme, Uri.UriSchemeHttps, StringComparison.Ordinal)
-            && pictureUri.Host.EndsWith(".googleusercontent.com", StringComparison.OrdinalIgnoreCase))
-        {
-            return true;
-        }
-
-        logger.LogWarning(
-            "Refusing to import Google photo for user {UserId}: URL is not a trusted Google host",
-            user.Id);
-        return false;
-    }
-
-    private async Task<byte[]?> FetchGoogleAvatarBytesAsync(Uri pictureUri, Guid userId, CancellationToken ct)
-    {
-        try
-        {
-            var httpClient = httpClientFactory.CreateClient(GoogleAvatarHttpClientName);
-            using var response = await httpClient.GetAsync(pictureUri, ct);
-            if (!response.IsSuccessStatusCode)
-            {
-                logger.LogWarning(
-                    "Google avatar fetch for user {UserId} returned {StatusCode}",
-                    userId, (int)response.StatusCode);
-                SetError(localizer["Profile_ImportGooglePhoto_FetchFailed"].Value);
-                return null;
-            }
-
-            var fetchedContentType = response.Content.Headers.ContentType?.MediaType ?? string.Empty;
-            if (!AllowedImageContentTypes.Contains(fetchedContentType))
-            {
-                logger.LogWarning(
-                    "Google avatar for user {UserId} returned unsupported content type {ContentType}",
-                    userId, fetchedContentType);
-                SetError(localizer["Profile_ImportGooglePhoto_InvalidFormat"].Value);
-                return null;
-            }
-
-            var rawBytes = await response.Content.ReadAsByteArrayAsync(ct);
-            if (rawBytes.Length == 0 || rawBytes.Length > MaxGooglePhotoDownloadBytes)
-            {
-                logger.LogWarning(
-                    "Google avatar for user {UserId} had invalid size {Bytes}", userId, rawBytes.Length);
-                SetError(localizer["Profile_ImportGooglePhoto_FetchFailed"].Value);
-                return null;
-            }
-
-            return rawBytes;
-        }
-        catch (HttpRequestException ex)
-        {
-            logger.LogWarning(ex, "Google avatar fetch failed for user {UserId}", userId);
-            SetError(localizer["Profile_ImportGooglePhoto_FetchFailed"].Value);
-            return null;
-        }
-        catch (TaskCanceledException ex) when (!ct.IsCancellationRequested)
-        {
-            logger.LogWarning(ex, "Google avatar fetch timed out for user {UserId}", userId);
-            SetError(localizer["Profile_ImportGooglePhoto_FetchFailed"].Value);
-            return null;
-        }
-    }
     // ─── View Another Profile ────────────────────────────────────────
 
     [HttpGet("{id:guid}")]
@@ -2518,4 +2388,3 @@ public class ProfileController(
     }
 
 }
-
