+++ src/Humans.Web/Extensions/Sections/ProfileSectionExtensions.cs
@@ -60,8 +60,6 @@ internal static class ProfileSectionExtensions
         // FullProfile cache retired — denormalized reads go through IUserService.GetUserInfoAsync.
         services.AddScoped<ProfilesProfileService>();
         services.AddScoped<IProfilePictureService>(sp => sp.GetRequiredService<ProfilesProfileService>());
-        services.AddScoped<IUserMerge>(sp => sp.GetRequiredService<ProfilesProfileService>());
-        services.AddScoped<IUserDataContributor>(sp => sp.GetRequiredService<ProfilesProfileService>());
 
         return services;
     }
