+++ src/Humans.Web/Models/ProfileViewModel.cs
@@ -15,7 +15,7 @@ public class ProfileViewModel
     public string? ProfilePictureUrl { get; set; }
 
     /// <summary>
-    /// Whether the profile has a custom uploaded picture (takes precedence over Google avatar).
+    /// Whether the profile has a custom uploaded picture.
     /// </summary>
     public bool HasCustomProfilePicture { get; set; }
 
@@ -23,13 +23,6 @@ public class ProfileViewModel
     /// URL to the custom profile picture endpoint (if uploaded).
     /// </summary>
     public string? CustomProfilePictureUrl { get; set; }
-
-    /// <summary>
-    /// True when the user can import a photo from Google — they signed in with Google,
-    /// we captured an avatar URL on <see cref="Humans.Domain.Entities.User.ProfilePictureUrl"/>,
-    /// and they don't yet have a custom uploaded picture. See issue #532.
-    /// </summary>
-    public bool CanImportGooglePicture { get; set; }
 
     [Required]
     [StringLength(100)]
@@ -245,7 +238,7 @@ public class ProfileViewModel
     public string? ApplicationRoleUnderstanding { get; set; }
 
     /// <summary>
-    /// The effective profile picture URL (custom upload takes priority over Google avatar).
+    /// The effective profile picture URL.
     /// </summary>
     public string? EffectiveProfilePictureUrl => HasCustomProfilePicture
         ? CustomProfilePictureUrl
