+++ src/Humans.Web/Resources/SharedResource.it.resx
@@ -2312,33 +2312,6 @@
   <data name="ProfileEdit_RemovePicture" xml:space="preserve">
     <value>Rimuovi foto personalizzata</value>
   </data>
-  <data name="ProfileEdit_ImportGooglePhoto" xml:space="preserve">
-    <value>Importa la mia foto di Google</value>
-  </data>
-  <data name="ProfileEdit_ImportGooglePhotoHelp" xml:space="preserve">
-    <value>Copia la foto del tuo account Google come foto del profilo. Potrai sostituirla in seguito caricandone una nuova.</value>
-  </data>
-  <data name="Profile_ImportGooglePhoto_Success" xml:space="preserve">
-    <value>La tua foto di Google è stata importata come foto del profilo.</value>
-  </data>
-  <data name="Profile_ImportGooglePhoto_Unavailable" xml:space="preserve">
-    <value>Non possiamo importare una foto di Google: il tuo account non è collegato a Google o nessun avatar è stato acquisito al momento dell'accesso.</value>
-  </data>
-  <data name="Profile_ImportGooglePhoto_NoProfile" xml:space="preserve">
-    <value>Completa prima la configurazione del tuo profilo, poi potrai importare la tua foto di Google.</value>
-  </data>
-  <data name="Profile_ImportGooglePhoto_AlreadyHasCustom" xml:space="preserve">
-    <value>Hai già una foto profilo personalizzata. Rimuovila prima per importare la tua foto di Google.</value>
-  </data>
-  <data name="Profile_ImportGooglePhoto_FetchFailed" xml:space="preserve">
-    <value>Non siamo riusciti a recuperare la tua foto di Google. Riprova più tardi o caricane una direttamente.</value>
-  </data>
-  <data name="Profile_ImportGooglePhoto_InvalidFormat" xml:space="preserve">
-    <value>La tua foto di Google non è stata elaborata: formato non supportato o risposta errata da Google.</value>
-  </data>
-  <data name="Profile_ImportGooglePhoto_NotGoogleUrl" xml:space="preserve">
-    <value>Non è stato possibile importare la tua foto di Google: l'URL dell'avatar memorizzato non sembra un indirizzo di Google. Caricane una direttamente.</value>
-  </data>
   <data name="ProfileEdit_DateOfBirthHelp" xml:space="preserve">
     <value>Visibile solo a te e al consiglio. Utilizzato per il calendario dei compleanni.</value>
   </data>
