+++ tests/Humans.Application.Tests/Controllers/ProfileControllerPopoverTests.cs
@@ -96,7 +96,6 @@ public class ProfileControllerPopoverTests
             Substitute.For<IApplicationDecisionService>(),
             Substitute.For<IAccountDeletionService>(),
             Substitute.For<IMembershipCalculator>(),
-            Substitute.For<IHttpClientFactory>(),
             signInManager,
             Options.Create(new GoogleWorkspaceOptions()));
